refactor: convert search engine property overrides to stateless converters
