+++ configuration/src/main/java/io/camunda/configuration/beans/SearchEngineIndexProperties.java
@@ -12,6 +12,6 @@ import io.camunda.search.schema.config.IndexConfiguration;
 /**
  * NOTE: Some of the fields of this object are overridden with values coming from the Unified
  * Configuration system, from the object
- * io.camunda.configuration.beanoverrides.SearchEngineIndexPropertiesOverride
+ * io.camunda.configuration.beanoverrides.SearchEngineIndexConverter
  */
 public class SearchEngineIndexProperties extends IndexConfiguration {}
