+++ configuration/src/main/java/io/camunda/configuration/beans/SearchEngineRetentionProperties.java
@@ -12,6 +12,6 @@ import io.camunda.search.schema.config.RetentionConfiguration;
 /**
  * NOTE: Some of the fields of this object are overridden with values coming from the Unified
  * Configuration system, from the object
- * io.camunda.configuration.beanoverrides.SearchEngineRetentionPropertiesOverride
+ * io.camunda.configuration.beanoverrides.SearchEngineRetentionConverter
  */
 public class SearchEngineRetentionProperties extends RetentionConfiguration {}
