+++ configuration/src/main/java/io/camunda/configuration/beans/SearchEngineSchemaManagerProperties.java
@@ -12,6 +12,6 @@ import io.camunda.search.schema.config.SchemaManagerConfiguration;
 /**
  * NOTE: Some of the fields of this object are overridden with values coming from the Unified
  * Configuration system, from the object
- * io.camunda.configuration.beanoverrides.SearchEngineSchemaManagerPropertiesOverride
+ * io.camunda.configuration.beanoverrides.SearchEngineSchemaManagerConverter
  */
 public class SearchEngineSchemaManagerProperties extends SchemaManagerConfiguration {}
