+++ configuration/src/test/java/io/camunda/configuration/SecondaryStorageElasticsearchTest.java
@@ -10,8 +10,8 @@ package io.camunda.configuration;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.camunda.configuration.beanoverrides.BrokerBasedPropertiesOverride;
-import io.camunda.configuration.beanoverrides.SearchEngineConnectPropertiesOverride;
-import io.camunda.configuration.beanoverrides.SearchEngineIndexPropertiesOverride;
+import io.camunda.configuration.beanoverrides.SearchEngineConnectConverter;
+import io.camunda.configuration.beanoverrides.SearchEngineIndexConverter;
 import io.camunda.configuration.beans.BrokerBasedProperties;
 import io.camunda.configuration.beans.SearchEngineConnectProperties;
 import io.camunda.configuration.beans.SearchEngineIndexProperties;
@@ -43,8 +43,6 @@ import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
   TasklistPropertiesOverride.class,
   OperatePropertiesOverride.class,
   BrokerBasedPropertiesOverride.class,
-  SearchEngineConnectPropertiesOverride.class,
-  SearchEngineIndexPropertiesOverride.class,
 })
 public class SecondaryStorageElasticsearchTest {
   private static final String EXPECTED_CLUSTER_NAME = "sample-cluster";
@@ -195,13 +193,12 @@ public class SecondaryStorageElasticsearchTest {
         @Autowired final OperateProperties operateProperties,
         @Autowired final TasklistProperties tasklistProperties,
         @Autowired final BrokerBasedProperties brokerBasedProperties,
-        @Autowired final SearchEngineConnectProperties searchEngineConnectProperties,
-        @Autowired final SearchEngineIndexProperties searchEngineIndexProperties) {
+        @Autowired final Camunda camunda) {
       this.operateProperties = operateProperties;
       this.tasklistProperties = tasklistProperties;
       this.brokerBasedProperties = brokerBasedProperties;
-      this.searchEngineConnectProperties = searchEngineConnectProperties;
-      this.searchEngineIndexProperties = searchEngineIndexProperties;
+      searchEngineConnectProperties = SearchEngineConnectConverter.convert(camunda);
+      searchEngineIndexProperties = SearchEngineIndexConverter.convert(camunda);
     }
 
     @Test
@@ -561,13 +558,12 @@ public class SecondaryStorageElasticsearchTest {
         @Autowired final OperateProperties operateProperties,
         @Autowired final TasklistProperties tasklistProperties,
         @Autowired final BrokerBasedProperties brokerBasedProperties,
-        @Autowired final SearchEngineConnectProperties searchEngineConnectProperties,
-        @Autowired final SearchEngineIndexProperties searchEngineIndexProperties) {
+        @Autowired final Camunda camunda) {
       this.operateProperties = operateProperties;
       this.tasklistProperties = tasklistProperties;
       this.brokerBasedProperties = brokerBasedProperties;
-      this.searchEngineConnectProperties = searchEngineConnectProperties;
-      this.searchEngineIndexProperties = searchEngineIndexProperties;
+      searchEngineConnectProperties = SearchEngineConnectConverter.convert(camunda);
+      searchEngineIndexProperties = SearchEngineIndexConverter.convert(camunda);
     }
 
     @Test
@@ -710,17 +706,14 @@ public class SecondaryStorageElasticsearchTest {
     final OperateProperties operateProperties;
     final TasklistProperties tasklistProperties;
     final BrokerBasedProperties brokerBasedProperties;
-    final SearchEngineConnectProperties searchEngineConnectProperties;
 
     ExporterTestWithoutArgs(
         @Autowired final OperateProperties operateProperties,
         @Autowired final TasklistProperties tasklistProperties,
-        @Autowired final BrokerBasedProperties brokerBasedProperties,
-        @Autowired final SearchEngineConnectProperties searchEngineConnectProperties) {
+        @Autowired final BrokerBasedProperties brokerBasedProperties) {
       this.operateProperties = operateProperties;
       this.tasklistProperties = tasklistProperties;
       this.brokerBasedProperties = brokerBasedProperties;
-      this.searchEngineConnectProperties = searchEngineConnectProperties;
     }
 
     // https://github.com/camunda/camunda/issues/37880
@@ -839,11 +832,11 @@ public class SecondaryStorageElasticsearchTest {
         @Autowired final OperateProperties operateProperties,
         @Autowired final TasklistProperties tasklistProperties,
         @Autowired final BrokerBasedProperties brokerBasedProperties,
-        @Autowired final SearchEngineConnectProperties searchEngineConnectProperties) {
+        @Autowired final Camunda camunda) {
       this.operateProperties = operateProperties;
       this.tasklistProperties = tasklistProperties;
       this.brokerBasedProperties = brokerBasedProperties;
-      this.searchEngineConnectProperties = searchEngineConnectProperties;
+      searchEngineConnectProperties = SearchEngineConnectConverter.convert(camunda);
     }
 
     @Test
@@ -908,11 +901,11 @@ public class SecondaryStorageElasticsearchTest {
         @Autowired final OperateProperties operateProperties,
         @Autowired final TasklistProperties tasklistProperties,
         @Autowired final BrokerBasedProperties brokerBasedProperties,
-        @Autowired final SearchEngineConnectProperties searchEngineConnectProperties) {
+        @Autowired final Camunda camunda) {
       this.operateProperties = operateProperties;
       this.tasklistProperties = tasklistProperties;
       this.brokerBasedProperties = brokerBasedProperties;
-      this.searchEngineConnectProperties = searchEngineConnectProperties;
+      searchEngineConnectProperties = SearchEngineConnectConverter.convert(camunda);
     }
 
     @Test
@@ -978,11 +971,11 @@ public class SecondaryStorageElasticsearchTest {
         @Autowired final OperateProperties operateProperties,
         @Autowired final TasklistProperties tasklistProperties,
         @Autowired final BrokerBasedProperties brokerBasedProperties,
-        @Autowired final SearchEngineConnectProperties searchEngineConnectProperties) {
+        @Autowired final Camunda camunda) {
       this.operateProperties = operateProperties;
       this.tasklistProperties = tasklistProperties;
       this.brokerBasedProperties = brokerBasedProperties;
-      this.searchEngineConnectProperties = searchEngineConnectProperties;
+      searchEngineConnectProperties = SearchEngineConnectConverter.convert(camunda);
     }
 
     @Test
@@ -1048,11 +1041,9 @@ public class SecondaryStorageElasticsearchTest {
     final Camunda camunda;
     final SearchEngineIndexProperties searchEngineIndexProperties;
 
-    WithDefaultValues(
-        @Autowired final Camunda camunda,
-        @Autowired final SearchEngineIndexProperties searchEngineIndexProperties) {
+    WithDefaultValues(@Autowired final Camunda camunda) {
       this.camunda = camunda;
-      this.searchEngineIndexProperties = searchEngineIndexProperties;
+      searchEngineIndexProperties = SearchEngineIndexConverter.convert(camunda);
     }
 
     @Test
