+++ operate/qa/integration-tests/src/test/java/io/camunda/operate/util/apps/modules/ModulesTestApplication.java
@@ -11,10 +11,6 @@ import io.camunda.configuration.UnifiedConfiguration;
 import io.camunda.configuration.UnifiedConfigurationHelper;
 import io.camunda.configuration.beanoverrides.GatewayBasedPropertiesOverride;
 import io.camunda.configuration.beanoverrides.GatewayRestPropertiesOverride;
-import io.camunda.configuration.beanoverrides.SearchEngineConnectPropertiesOverride;
-import io.camunda.configuration.beanoverrides.SearchEngineIndexPropertiesOverride;
-import io.camunda.configuration.beanoverrides.SearchEngineRetentionPropertiesOverride;
-import io.camunda.configuration.beanoverrides.SearchEngineSchemaManagerPropertiesOverride;
 import io.camunda.operate.OperateModuleConfiguration;
 import io.camunda.operate.OperatePropertiesOverride;
 import io.camunda.operate.util.TestApplication;
@@ -53,10 +49,6 @@ import org.springframework.context.annotation.Import;
   // Unified Configuration classes
   UnifiedConfigurationHelper.class,
   UnifiedConfiguration.class,
-  SearchEngineConnectPropertiesOverride.class,
-  SearchEngineIndexPropertiesOverride.class,
-  SearchEngineRetentionPropertiesOverride.class,
-  SearchEngineSchemaManagerPropertiesOverride.class,
   OperatePropertiesOverride.class,
   GatewayBasedPropertiesOverride.class,
   GatewayRestPropertiesOverride.class,
